изменения в модели - поля долгота и широта исправлены из int  в double. Внесены изменения в GradutesCntroller Index.html добавлены изображения в папку imege добавлен скрипт в пакпку js

+++ src/Board/Data/Graduate.cs
@@ -23,9 +23,9 @@ namespace Board.Data
 
         public string History { get; set; }
 
-        public int Latitude { get; set; }
+        public double Latitude { get; set; }
 
-        public int Longitude { get; set; }
+        public double Longitude { get; set; }
 
     }
 }
